refactor: move Money class into product package
