+++ src/test/kotlin/com/kroffle/knitting/helper/extension/Money.kt
@@ -1,6 +1,6 @@
 package com.kroffle.knitting.helper.extension
 
-import com.kroffle.knitting.domain.value.Money
+import com.kroffle.knitting.domain.product.value.Money
 
 fun Money.like(other: Money): Boolean {
     if (this === other) return true
